+++ lib/storage_api/src/batch.rs
@@ -1,3 +1,4 @@
+use crate::ReadFinality;
 use alloy::primitives::BlockNumber;
 
 #[async_trait::async_trait]
@@ -6,6 +7,8 @@ pub trait ReadBatch: Send + Sync + 'static {
     async fn get_batch_by_block_number(
         &self,
         block_number: BlockNumber,
+        // todo(#205): remove/refactor once batch storage cache is in place
+        finality: &dyn ReadFinality,
     ) -> anyhow::Result<Option<u64>>;
 
     // todo: return `BatchMetadata` once it is moved to `types` crate