+++ lib/storage_api/src/finality.rs
@@ -1,8 +1,11 @@
 use crate::FinalityStatus;
+use tokio::sync::watch::Receiver;
 
 pub trait ReadFinality: Send + Sync + 'static {
     /// Get latest finality status.
     fn get_finality_status(&self) -> FinalityStatus;
+
+    fn subscribe(&self) -> Receiver<FinalityStatus>;
 }
 
 pub trait WriteFinality: ReadFinality {