+++ lib/storage_api/src/model.rs
@@ -81,7 +81,9 @@ impl ReplayRecord {
 #[derive(Clone, Debug)]
 pub struct FinalityStatus {
     pub last_committed_block: u64,
+    pub last_committed_batch: u64,
     pub last_executed_block: u64,
+    pub last_executed_batch: u64,
 }
 
 #[cfg(test)]